% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamaddexp.R
\name{iscamaddexp}
\alias{iscamaddexp}
\title{iscamaddexp Function}
\usage{
iscamaddexp(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays an exponential density function with lambda = 1/mean.
}
\examples{
iscamaddexp(x)
}
\keyword{exponential}
