% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamaddlnorm.R
\name{iscamaddlnorm}
\alias{iscamaddlnorm}
\title{iscamaddlnorm Function}
\usage{
iscamaddlnorm(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays a log normal density function.
}
\examples{
iscamaddlnorm(x)
}
\keyword{lognormal}
