% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamaddnorm.R
\name{iscamaddnorm}
\alias{iscamaddnorm}
\title{iscamaddnorm Function}
\usage{
iscamaddnorm(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\description{
This function creates a histogram of the inputted variable and overlays a normal density function.
}
\examples{
iscamaddnorm(x)
}
\keyword{normal}
