% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamaddt.R
\name{iscamaddt}
\alias{iscamaddt}
\title{iscamaddt Function}
\usage{
iscamaddt(x, df)
}
\arguments{
\item{x}{a vector of numeric values.}

\item{df}{degrees of freedom}
}
\description{
This function creates a histogram of the inputted variable and overlays a t density function with df degrees of freedom.
}
\examples{
iscamaddt(x, 3)
}
\keyword{t}
