% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscambinom.R
\name{iscambinom}
\alias{iscambinom}
\title{iscambinom Function}
\usage{
iscambinom(k, prob, n, pi, lower.tail)
}
\arguments{
\item{k}{number of successes}

\item{prob}{probability of interest}

\item{n}{number of trials (zero or more)}

\item{pi}{probability of success on each trial}

\item{lower.tail}{logical; if TRUE, probabilities are  P[X ??? k], otherwise, P[X > k]}
}
\description{
This function calculates either a tail probability or the inverse cumulative 
probability depending on whether k or prob is passed a question mark.
}
\examples{
iscambinom(?, .50, 20, lower.tail = TRUE)
iscambinom(10, ?, 20, lower.tail = TRUE)
}
\keyword{binomial}
