% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscambinompower.R
\name{iscambinompower}
\alias{iscambinompower}
\title{iscambinompower Function}
\usage{
iscambinompower(LOS, n, prob1, alternative, prob2 = NULL)
}
\arguments{
\item{LOS}{level of significance}

\item{n}{number of trials (zero or more)}

\item{prob1}{probability of success on each trial}

\item{alternative}{allows you to specify whether you want to find the probability 
"less" or "greater" or a symmetric "two.sided" probability}

\item{prob2}{NULL or a second probability}
}
\description{
This function determines the rejection region corresponding to the level of 
significance and the first probability and shows the second distribution shading 
its corresponding region.
}
\examples{
iscambinompower(.05, 30, 0.5, alternative = "above")
iscambinompower(.10, 55, 0.10, alternative = "below")
iscambinompower(.05, 144, 0.40, alternative = "two.sided")
}
\keyword{binomial}
\keyword{power}
