% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscambinomprob.R
\name{iscambinomprob}
\alias{iscambinomprob}
\title{iscambinomprob Function}
\usage{
iscambinomprob(k, n, prob, lower.tail)
}
\arguments{
\item{k}{number of successes of interest (must be an integer)}

\item{n}{number of trials (zero or more)}

\item{prob}{probability of success on each trial}

\item{lower.tail}{logical; a Boolean for finding the probability above (FALSE) or below (TRUE) the inputted value (inclusive)}
}
\description{
This function calculates tail probabilities from the binomial distribution.
}
\examples{
iscambinomprob(20, 30, 0.5, lower.tail = TRUE)
iscambinomprob(10, 55, 0.10, lower.tail = FALSE)
}
\keyword{binomial}
