% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscambinomtest.R
\name{iscambinomtest}
\alias{iscambinomtest}
\title{iscambinomtest Function}
\usage{
iscambinomtest(observed, n, hypothesized = NULL, alternative,
  conf.level = NULL)
}
\arguments{
\item{observed}{number of successes of interest or sample proportion (assumed if value less than one)}

\item{n}{sample size}

\item{hypothesized}{NULL or probability of success on each trial}

\item{alternative}{form of alternative "less", "greater", or "two.sided"}

\item{conf.level}{NULL or a confidence level (one or more values) for a two-sided confidence interval}
}
\description{
This function performs an exact binomial test and graphs the binomial distribution and/or binomial confidence interval.
}
\examples{
iscambinomtest(20, 35, alternative = "less")
iscambinomtest(10, 40, hypothesized = .50, alternative = "two.sided", conf.level = 0.90)
}
\keyword{binomial}
