% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamboxplot.R
\name{iscamboxplot}
\alias{iscamboxplot}
\title{iscamboxplot Function}
\usage{
iscamboxplot(x, explanatory = NULL, names = NULL)
}
\arguments{
\item{x}{a numeric vector or a single list containing numeric vectors form which the boxplots are to be produced}

\item{explanatory}{NULL or a second, categorical grouping variable}

\item{names}{used to specify the horizontal and vertical axis labels respectively}
}
\description{
This function displays horizontal boxplot(s) utilizing quartiles instead of hinges to match the summary statistics.
}
\examples{
iscamboxplot(x)
iscamboxplot(mtcars$mpg, explanatory = mtcars$vs, names = c("mpg", "vs"))
}
\keyword{boxplot}
