% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamchisqprob.R
\name{iscamchisqprob}
\alias{iscamchisqprob}
\title{iscamchisqprob Function}
\usage{
iscamchisqprob(xval, df)
}
\arguments{
\item{xval}{x value}

\item{df}{degrees of freedom}
}
\description{
This function calculates the upper tail probability for the chi-square distribution
}
\examples{
iscamchisqprob(10, 4)
}
\keyword{chi}
\keyword{square}
