% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamdotplot.R
\name{iscamdotplot}
\alias{iscamdotplot}
\title{iscamdotplot Function}
\usage{
iscamdotplot(response, explanatory = NULL, names = NULL)
}
\arguments{
\item{response}{a numeric vector or a single list containing numeric vectors from which the dotplots are to be produced}

\item{explanatory}{grouping variable}

\item{names}{used to specify the horizontal and vertical axis labels respectively}
}
\description{
This function creates horizontal dotplots.If no explanatory variable is specified, just one plot. If an explanatory variable is specified, parallel dotplots will be created.
}
\examples{
iscamdotplot(x)
iscamdotplot(mtcars$mpg, explanatory = mtcars$vs, names = c("mpg", "vs"))
}
\keyword{dotplot}
