% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscaminvnorm.R
\name{iscaminvnorm}
\alias{iscaminvnorm}
\title{iscaminvnorm Function}
\usage{
iscaminvnorm(prob1, mean = 0, sd = 1, direction)
}
\arguments{
\item{prob1}{desired probability}

\item{mean}{default = 0}

\item{sd}{default = 1}

\item{direction}{Specify whether you want this area to be "above", "below", or "outside" or "between"}
}
\description{
This function calculates the normal quantile of a specified probability.
}
\examples{
iscaminvnorm(.55, 15, 2, direction = "above")
}
\keyword{inverse}
\keyword{normal}
\keyword{probability}
