% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscaminvt.R
\name{iscaminvt}
\alias{iscaminvt}
\title{iscaminvt Function}
\usage{
iscaminvt(prob1, df, direction)
}
\arguments{
\item{prob1}{desired probability}

\item{df}{degrees of freedom}

\item{direction}{Specify whether you want this area to be "above", "below", or "outside" or "between"}
}
\description{
This function calculates the t quantile of a specified probability.
}
\examples{
iscaminvt(.55, 5, direction = "outside")
}
\keyword{inverse}
\keyword{probability}
\keyword{t}
