% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamnormpower.R
\name{iscamnormpower}
\alias{iscamnormpower}
\title{iscamnormpower Function}
\usage{
iscamnormpower(LOS, n, prob1, alternative, prob2)
}
\arguments{
\item{LOS}{level of significance}

\item{n}{sample size}

\item{prob1}{probability of success}

\item{alternative}{can be "less", "greater", or "two.sided"}

\item{prob2}{second probability of success}
}
\description{
This function determines the rejection region corresponding to the level of significance and the first probability and shows the second distribution shading its corresponding region.
}
\examples{
iscamnormpower(.05, 20, .5, "less", .7)
}
\keyword{normal}
\keyword{power}
