% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscamsummary.R
\name{iscamsummary}
\alias{iscamsummary}
\title{iscamsummary Function}
\usage{
iscamsummary(x, explanatory = NULL, digits = 3)
}
\arguments{
\item{x}{quantitative variable}

\item{explanatory}{optional, categorical grouping variable}

\item{digits}{number of tick marks on x axis}
}
\description{
This function calculates the five number summary, mean, and standard deviation of the quantitative variable x.
}
\examples{
iscamsummary(y, x)
}
\keyword{summary}
